Created a new file "BooleanReflection.v" and added basic definitions. Undid earlier suggested idea of a "Complementary Prop", two props A, B with A + B and A -> ~B, i.eb. A = n <= m and B = n > m. I am not convinced yet that this will be useful enough to be worth adding.
Created a new file "BooleanReflection.v" and added basic definitions.

Created a new file SProp.v and added basic definitions.

Changed definition of Nat.lt so that it does not take a universe as parameter, instead it is hardcoded to Type0.
Changed files under Classes/implementations and Classes/theory so the change to Nat.lt is not breaking.